pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "PyStegano"
version = "0.1.0"
authors = [
  { name="gerardo burgos ", email="gerardoburgos1232@gmail.com" },
]
description = "Una librería simple y enfocada para esteganografía de imágenes."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Security",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "Pillow>=9.0.0"
]

[project.urls]
"Homepage" = "https://github.com/gerarb1/PyStegano.git"
"Bug Tracker" = "https://github.com/gerarb1"

[tool.setuptools.packages.find]
where = ["."]
include = ["PyStegano*"]
namespaces = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
